feat(desktop): render trigger context as a typed card in agent sessions
Replace the raw <trigger_context>/<trigger_instructions> text shown at the
top of trigger-created sessions with a friendly, per-trigger-type card.

- Co-locate all per-trigger-type frontend logic (option labels, defaults,
  source summaries, configuration forms, activity-route derivation, and the
  new TriggerContextView) in pages/agents/trigger-types.tsx so each trigger
  type stays in sync across the trigger forms and the session UI.
- Hide the model-facing trigger blocks from the first message bubble; the
  exact markdown remains available via the raw-markdown dialog.
- Link the trigger name to the trigger page and the activity summary to the
  exact comment, mention source, or document version that fired the trigger,
  reusing the activity feed's getEventRoute (now exported from @shm/ui/feed).
- Add collapsible "Activity details" and "Trigger instructions" sections.
