+++ frontend/apps/desktop/src/pages/agents/detail.tsx
@@ -31,13 +31,9 @@ import {useSelectedAccountId} from '@/selected-account'
 import {useClickNavigate, useNavigate} from '@/utils/useNavigate'
 import {markdownBlockNodesToHMBlockNodes, parseMarkdown} from '@seed-hypermedia/client'
 import type {HMBlockNode} from '@seed-hypermedia/client/hm-types'
-import {useSearch} from '@shm/shared/models/search'
-import {abbreviateUid} from '@shm/shared/utils/abbreviate'
 import {formattedDateMedium} from '@shm/shared/utils/date'
-import {hmId, packHmId} from '@shm/shared/utils/entity-id-url'
 import {useNavRoute} from '@shm/shared/utils/navigation'
 import {Button} from '@shm/ui/button'
-import {AccountSearchInput, type SearchResult} from '@shm/ui/collaborators-page'
 import {
   AlertDialogAction,
   AlertDialogCancel,
@@ -56,6 +52,7 @@ import {KeyRound, Plus, Trash2} from 'lucide-react'
 import React, {useCallback, useEffect, useMemo, useRef, useState} from 'react'
 import {seedToolRegistry} from '../../../../../../agents/protocol/src/tool-registry'
 import {AGENT_READ_TOOL_GROUP} from './agent-tools'
+import {TriggerSourceFields, summarizeTriggerSource} from './trigger-types'
 import {AddModelProviderDialog, EditAgentNameDialog, type AgentAccountRenameStatus} from './dialogs'
 import {AgentHeader, AgentSubpageHeader, type AgentPageTab} from './header'
 import {ModelSelect} from './model-select'
@@ -1148,7 +1145,7 @@ function AgentTriggersTab({
             </SizableText>
           </div>
           <SizableText size="sm" color="muted">
-            {triggerSourceSummary(item.source)}
+            {summarizeTriggerSource(item.source)}
           </SizableText>
           <SizableText size="xs" color="muted">
             Updated {new Date(item.updatedAt).toLocaleString()}
@@ -1184,7 +1181,7 @@ function CreateAgentTriggerDialog({
   const [enabled, setEnabled] = useState(true)
   const [source, setSource] = useState<AgentTriggerSource>({type: 'document-comment', resource: ''})
   const [prompt, setPrompt] = useState<HMBlockNode[]>(() =>
-    agentPromptToBlocks('Read the related Seed context and summarize what needs attention.'),
+    agentPromptToBlocks('Respond to the mention, performing the action requested.'),
   )
 
   async function handleCreateTrigger() {
@@ -1237,420 +1234,6 @@ function CreateAgentTriggerDialog({
       </div>
     </div>
   )
-}
-
-function TriggerSourceFields({
-  source,
-  onChange,
-}: {
-  source: AgentTriggerSource
-  onChange: (source: AgentTriggerSource) => void
-}) {
-  return (
-    <div className="grid gap-3">
-      <label className="flex flex-col gap-1">
-        <SizableText size="sm" weight="bold">
-          Trigger Session on:
-        </SizableText>
-        <select
-          className="border-border bg-input rounded-md border px-3 py-2 text-sm"
-          value={source.type}
-          onChange={(event) => onChange(defaultSourceForType(event.target.value as AgentTriggerSource['type']))}
-        >
-          <option value="document-comment">Comment in a document</option>
-          <option value="user-mention">User mention</option>
-          <option value="site-update">Site update</option>
-          <option value="schedule">Schedule</option>
-        </select>
-      </label>
-      {source.type === 'document-comment' ? (
-        <div className="grid gap-3 md:grid-cols-2">
-          <DocumentAutocompleteField
-            label="Document"
-            value={source.resource}
-            onChange={(value) => onChange({...source, resource: value})}
-            placeholder="Search documents or enter hm:// URL"
-          />
-          <label className="flex flex-col gap-1">
-            <SizableText size="sm" weight="bold">
-              Author filter
-            </SizableText>
-            <Input
-              value={source.author || ''}
-              onChange={(event) => onChange({...source, author: event.target.value || undefined})}
-              placeholder="optional account ID"
-            />
-          </label>
-        </div>
-      ) : null}
-      {source.type === 'user-mention' ? (
-        <div className="grid gap-3">
-          <MentionedAccountsField
-            accounts={mentionedAccountsOf(source)}
-            onChange={(accounts) => onChange({...source, mentionedAccounts: accounts})}
-          />
-        </div>
-      ) : null}
-      {source.type === 'site-update' ? (
-        <div className="grid gap-3 md:grid-cols-2">
-          <AccountAutocompleteField
-            label="Resource/site prefix"
-            value={source.resourcePrefix}
-            onChange={(value) => onChange({...source, resourcePrefix: value})}
-            placeholder="Search site/account or enter hm:// prefix"
-            valueFormat="hm-url"
-          />
-          <label className="flex flex-col gap-1">
-            <SizableText size="sm" weight="bold">
-              Event types
-            </SizableText>
-            <Input
-              value={(source.eventTypes || []).join(', ')}
-              onChange={(event) =>
-                onChange({
-                  ...source,
-                  eventTypes: event.target.value
-                    .split(',')
-                    .map((value) => value.trim())
-                    .filter(Boolean),
-                })
-              }
-              placeholder="doc-update, comment"
-            />
-          </label>
-        </div>
-      ) : null}
-      {source.type === 'schedule' ? <ScheduleTriggerFields source={source} onChange={onChange} /> : null}
-    </div>
-  )
-}
-
-function ScheduleTriggerFields({
-  source,
-  onChange,
-}: {
-  source: Extract<AgentTriggerSource, {type: 'schedule'}>
-  onChange: (source: AgentTriggerSource) => void
-}) {
-  const schedule = source.schedule
-  const timezone = Intl.DateTimeFormat().resolvedOptions().timeZone || 'UTC'
-  const setSchedule = (next: Extract<AgentTriggerSource, {type: 'schedule'}>['schedule']) =>
-    onChange({type: 'schedule', schedule: next})
-  return (
-    <div className="grid gap-3">
-      <label className="flex flex-col gap-1">
-        <SizableText size="sm" weight="bold">
-          Schedule mode
-        </SizableText>
-        <select
-          className="border-border bg-input rounded-md border px-3 py-2 text-sm"
-          value={schedule.kind}
-          onChange={(event) => {
-            const kind = event.target.value
-            if (kind === 'weekly') setSchedule({kind, daysOfWeek: [1, 2, 3, 4, 5], timeOfDay: '09:00', timezone})
-            else if (kind === 'once') setSchedule({kind, runAt: Date.now() + 60 * 60 * 1000, timezone})
-            else setSchedule({kind: 'interval', every: 1, unit: 'hours'})
-          }}
-        >
-          <option value="interval">Every interval</option>
-          <option value="weekly">Days of week</option>
-          <option value="once">One time</option>
-        </select>
-      </label>
-      {schedule.kind === 'interval' ? (
-        <div className="grid gap-3 md:grid-cols-2">
-          <label className="flex flex-col gap-1">
-            <SizableText size="sm" weight="bold">
-              Every
-            </SizableText>
-            <Input
-              type="number"
-              min={1}
-              value={schedule.every}
-              onChange={(event) => setSchedule({...schedule, every: Number(event.target.value) || 1})}
-            />
-          </label>
-          <label className="flex flex-col gap-1">
-            <SizableText size="sm" weight="bold">
-              Unit
-            </SizableText>
-            <select
-              className="border-border bg-input rounded-md border px-3 py-2 text-sm"
-              value={schedule.unit}
-              onChange={(event) => setSchedule({...schedule, unit: event.target.value as 'minutes' | 'hours'})}
-            >
-              <option value="minutes">Minutes</option>
-              <option value="hours">Hours</option>
-            </select>
-          </label>
-        </div>
-      ) : null}
-      {schedule.kind === 'weekly' ? (
-        <div className="grid gap-3">
-          <div className="flex flex-wrap gap-2">
-            {[
-              ['Mon', 1],
-              ['Tue', 2],
-              ['Wed', 3],
-              ['Thu', 4],
-              ['Fri', 5],
-              ['Sat', 6],
-              ['Sun', 0],
-            ].map(([day, dayIndex]) => (
-              <label key={day} className="border-border flex items-center gap-2 rounded-md border px-3 py-2 text-sm">
-                <input
-                  type="checkbox"
-                  checked={schedule.daysOfWeek.includes(dayIndex as number)}
-                  onChange={(event) => {
-                    const dayNumber = dayIndex as number
-                    const daysOfWeek = event.target.checked
-                      ? [...schedule.daysOfWeek, dayNumber].sort()
-                      : schedule.daysOfWeek.filter((item) => item !== dayNumber)
-                    setSchedule({...schedule, daysOfWeek})
-                  }}
-                />
-                {day}
-              </label>
-            ))}
-          </div>
-          <div className="grid gap-3 md:grid-cols-2">
-            <label className="flex flex-col gap-1">
-              <SizableText size="sm" weight="bold">
-                Time of day
-              </SizableText>
-              <Input
-                type="time"
-                value={schedule.timeOfDay}
-                onChange={(event) => setSchedule({...schedule, timeOfDay: event.target.value})}
-              />
-            </label>
-            <label className="flex flex-col gap-1">
-              <SizableText size="sm" weight="bold">
-                Timezone
-              </SizableText>
-              <Input
-                value={schedule.timezone}
-                onChange={(event) => setSchedule({...schedule, timezone: event.target.value})}
-              />
-            </label>
-          </div>
-        </div>
-      ) : null}
-      {schedule.kind === 'once' ? (
-        <div className="grid gap-3 md:grid-cols-2">
-          <label className="flex flex-col gap-1">
-            <SizableText size="sm" weight="bold">
-              Date and time
-            </SizableText>
-            <Input
-              type="datetime-local"
-              value={dateTimeLocalValue(schedule.runAt)}
-              onChange={(event) => setSchedule({...schedule, runAt: new Date(event.target.value).getTime(), timezone})}
-            />
-          </label>
-          <label className="flex flex-col gap-1">
-            <SizableText size="sm" weight="bold">
-              Timezone
-            </SizableText>
-            <Input
-              value={schedule.timezone || timezone}
-              onChange={(event) => setSchedule({...schedule, timezone: event.target.value})}
-            />
-          </label>
-        </div>
-      ) : null}
-    </div>
-  )
-}
-
-function DocumentAutocompleteField({
-  label,
-  value,
-  onChange,
-  placeholder,
-}: {
-  label: string
-  value: string
-  onChange: (value: string) => void
-  placeholder: string
-}) {
-  const [focused, setFocused] = useState(false)
-  const search = useSearch(value, {
-    enabled: focused && value.trim().length > 0,
-    pageSize: 12,
-  })
-  const documents = useMemo(
-    () => (search.data?.entities || []).filter((item) => item.type === 'document').slice(0, 8),
-    [search.data?.entities],
-  )
-
-  return (
-    <label className="relative flex flex-col gap-1">
-      <SizableText size="sm" weight="bold">
-        {label}
-      </SizableText>
-      <Input
-        value={value}
-        onFocus={() => setFocused(true)}
-        onBlur={() => window.setTimeout(() => setFocused(false), 120)}
-        onChange={(event) => onChange(event.target.value)}
-        placeholder={placeholder}
-      />
-      {focused && documents.length ? (
-        <div className="border-border bg-popover absolute top-full right-0 left-0 z-20 mt-1 max-h-64 overflow-auto rounded-md border p-1 shadow-lg">
-          {documents.map((document) => {
-            const nextValue = packHmId(document.id)
-            return (
-              <button
-                key={document.id.id}
-                type="button"
-                className="hover:bg-muted flex w-full flex-col rounded px-2 py-2 text-left"
-                onMouseDown={(event) => event.preventDefault()}
-                onClick={() => {
-                  onChange(nextValue)
-                  setFocused(false)
-                }}
-              >
-                <SizableText size="sm" weight="bold" className="truncate">
-                  {document.title || nextValue}
-                </SizableText>
-                <SizableText size="xs" color="muted" className="truncate font-mono">
-                  {nextValue}
-                </SizableText>
-              </button>
-            )
-          })}
-        </div>
-      ) : null}
-    </label>
-  )
-}
-
-function AccountAutocompleteField({
-  label,
-  value,
-  onChange,
-  placeholder,
-  valueFormat,
-}: {
-  label: string
-  value: string
-  onChange: (value: string) => void
-  placeholder: string
-  valueFormat: 'uid' | 'hm-url'
-}) {
-  const [focused, setFocused] = useState(false)
-  const search = useSearch(value, {
-    enabled: focused && value.trim().length > 0,
-    pageSize: 12,
-  })
-  const accounts = useMemo(
-    () => (search.data?.entities || []).filter((item) => item.type === 'contact' || !item.id.path?.length).slice(0, 8),
-    [search.data?.entities],
-  )
-
-  return (
-    <label className="relative flex flex-col gap-1">
-      <SizableText size="sm" weight="bold">
-        {label}
-      </SizableText>
-      <Input
-        value={value}
-        onFocus={() => setFocused(true)}
-        onBlur={() => window.setTimeout(() => setFocused(false), 120)}
-        onChange={(event) => onChange(event.target.value)}
-        placeholder={placeholder}
-      />
-      {focused && accounts.length ? (
-        <div className="border-border bg-popover absolute top-full right-0 left-0 z-20 mt-1 max-h-64 overflow-auto rounded-md border p-1 shadow-lg">
-          {accounts.map((account) => {
-            const nextValue = valueFormat === 'hm-url' ? `hm://${account.id.uid}` : account.id.uid
-            return (
-              <button
-                key={`${account.id.id}:${account.type}`}
-                type="button"
-                className="hover:bg-muted flex w-full flex-col rounded px-2 py-2 text-left"
-                onMouseDown={(event) => event.preventDefault()}
-                onClick={() => {
-                  onChange(nextValue)
-                  setFocused(false)
-                }}
-              >
-                <SizableText size="sm" weight="bold" className="truncate">
-                  {account.title || account.id.uid}
-                </SizableText>
-                <SizableText size="xs" color="muted" className="truncate font-mono">
-                  {nextValue}
-                </SizableText>
-              </button>
-            )
-          })}
-        </div>
-      ) : null}
-    </label>
-  )
-}
-
-/** Reads the mentioned account list, tolerating legacy triggers that stored a single `mentionedAccount`. */
-function mentionedAccountsOf(source: Extract<AgentTriggerSource, {type: 'user-mention'}>): string[] {
-  const legacy = (source as {mentionedAccount?: string}).mentionedAccount
-  return source.mentionedAccounts ?? (legacy ? [legacy] : [])
-}
-
-function MentionedAccountsField({
-  accounts,
-  onChange,
-}: {
-  accounts: string[]
-  onChange: (accounts: string[]) => void
-}) {
-  const accountsKey = accounts.join('|')
-  const values = useMemo<SearchResult[]>(
-    () => accounts.map((uid) => ({id: hmId(uid), label: abbreviateUid(uid), unresolved: true})),
-    // accountsKey captures the contents of `accounts` for memoization
-    [accountsKey],
-  )
-  return (
-    <div className="flex flex-col gap-1">
-      <SizableText size="sm" weight="bold">
-        Mentioned accounts
-      </SizableText>
-      <div className="border-border bg-input flex min-h-9 items-center overflow-hidden rounded-md border">
-        <AccountSearchInput
-          label="Mentioned accounts"
-          placeholder="Search or paste accounts"
-          values={values}
-          onValuesChange={(next) => onChange(next.map((value) => value.id.uid))}
-        />
-      </div>
-    </div>
-  )
-}
-
-function defaultSourceForType(type: AgentTriggerSource['type']): AgentTriggerSource {
-  if (type === 'user-mention') return {type, mentionedAccounts: []}
-  if (type === 'site-update') return {type, resourcePrefix: '', eventTypes: ['doc-update', 'comment']}
-  if (type === 'schedule') return {type, schedule: {kind: 'interval', every: 1, unit: 'hours'}}
-  return {type: 'document-comment', resource: ''}
-}
-
-function triggerSourceSummary(source: AgentTriggerSource): string {
-  if (source.type === 'document-comment') {
-    return `Comment in ${source.resource}${source.author ? ` by ${source.author}` : ''}`
-  }
-  if (source.type === 'user-mention') {
-    const accounts = mentionedAccountsOf(source)
-    const mention = accounts.length ? accounts.map(abbreviateUid).join(', ') : 'anyone'
-    return `Mention of ${mention}${source.resourcePrefix ? ` in ${source.resourcePrefix}` : ''}`
-  }
-  if (source.type === 'site-update') {
-    return `Update in ${source.resourcePrefix}${source.eventTypes?.length ? ` (${source.eventTypes.join(', ')})` : ''}`
-  }
-  if (source.schedule.kind === 'interval') return `Every ${source.schedule.every} ${source.schedule.unit}`
-  if (source.schedule.kind === 'once') return `Once at ${formattedDateMedium(new Date(source.schedule.runAt))}`
-  return `${source.schedule.daysOfWeek.map(dayName).join(', ')} at ${source.schedule.timeOfDay} ${
-    source.schedule.timezone
-  }`
 }
 
 function nextScheduleFire(input: {
@@ -1737,17 +1320,6 @@ function zonedParts(
     minute: Number(values.minute),
     weekday: ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'].indexOf((values.weekday || 'Sun').slice(0, 3)),
   }
-}
-
-function dateTimeLocalValue(ms: number): string {
-  if (!Number.isFinite(ms)) return ''
-  const date = new Date(ms)
-  const offsetMs = date.getTimezoneOffset() * 60 * 1000
-  return new Date(date.getTime() - offsetMs).toISOString().slice(0, 16)
-}
-
-function dayName(day: number): string {
-  return ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'][day] || String(day)
 }
 
 function StartSessionInput({
